partners section update admin panel for title and desc

+++ database/migrations/2025_09_11_151757_add_logo_and_copyright_to_site_settings_table.php
@@ -6,18 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-public function up(): void
-{
-    Schema::table('site_settings', function (Blueprint $table) {
-        $table->string('logo')->nullable()->after('id'); // only add logo
-    });
-}
+    public function up(): void
+    {
+        Schema::table('site_settings', function (Blueprint $table) {
+            if (!Schema::hasColumn('site_settings', 'logo')) {
+                $table->string('logo')->nullable()->after('id');
+            }
+        });
+    }
 
-public function down(): void
-{
-    Schema::table('site_settings', function (Blueprint $table) {
-        $table->dropColumn('logo');
-    });
-}
-
+    public function down(): void
+    {
+        Schema::table('site_settings', function (Blueprint $table) {
+            if (Schema::hasColumn('site_settings', 'logo')) {
+                $table->dropColumn('logo');
+            }
+        });
+    }
 };
